requirements.txt
matplotlib
python-dotenv
pytz